lychee-core/pyproject.toml
[project]
name = "lychee-core"
dynamic = ["version"]
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Marcus Figueiredo", email = "figueiredo@protonmail.com" }]
requires-python = ">=3.11"
dependencies = [
    "jsonschema>=4.25.1",
    "psutil>=7.0.0",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "watchfiles>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ['src/lychee']

[tool.hatch.version]
path = 'src/lychee/core/__init__.py'
